check deps during bootup
